fixed "fail walking" bug
when you fail a trick and quickly swap to walking mode, there is a
change that you can get stuck in place. And even if you manage to
unstuck yourself somehow, you can't push with skate. I believe that I
fixed it with this commit, since I can't recreate that bug?